Filter attachment meta data for Tachyon.
This fixes two issues.

In Tachyon, image sizes created after upload are not included in the srcset as they are not included in the image meta data stored by WordPress.

In `wp_calculate_image_srcset()` WordPress does not run the srcset filter if the image metadata does not include resized images. For users of Tachyon who disable resize on upload, no resized images exist.