Support for Gutenberg style images in content
Gutenberg changed how it inserts images into posts quite a lot. Firstly, it now fetches the image URLs aver the REST API, so we were incorrectly inserting Tachyon image URLs into post content, which we don't want to do. It's always been intended that by default, you don't get Tachyon URLs in the database.

Once I fixed up this, I also noticed we have numerous issues in the `the_content` image url replacements, again because they are different markup with the Gutenberg editor. For example, the `size` class is no longer added, so it's a lot more difficult to generate the correct url.

Also, custom hooks on `tachyon_args` and similar are not working as expected, because the context is different. Previously image URLs were generated via `image_downsize` in Tachyon, because that was inserted via the REST API. That means when we have to do a direct WP image url -> Tachyon URL (which is what the `the_content` filter does), there are more edge cases to account for.

This commit cleans up these edge cases, ultimately to restore previous pre-Gutenberg behaviour.