Ensure attachment meta contains width/height before resizing
It's possible for `wp_get_attachment_metadata` to return false, or to
return a metadata object which doesn't have width and height propoerties
defined. In those cases, this code would previously have thrown a
warning. This PR adds some defensive coding to prevent PHP warning, in
a similar way to how attachment meta is dealt with elsewhere in this
file.