Handle missing metadata
In certain situations, especially with migrated legacy content, an image
URL may be sent to tachyon for an image which doesn't exist on disk and
may not have any stored metadata. In this case, because we never check
that the meta array is well formed and didn't return as "false", several
log warnings will be emitted:

```
Trying to access array offset on value of type bool | …achyon-plugin/inc/class-tachyon.php#631
Trying to access array offset on value of type bool | …achyon-plugin/inc/class-tachyon.php#632
```

Almost all of our calculations in this `if` block assume that meta
exists and ws returned as an array. If that's not true, we should move
on to compute entirely based on the provided arguments.